Shrink AST size, decrease allocated memory while idle (#251)
* shrink ast size, decrease allocated memory while idle

* fix test compilation issue

* return the visit objects count from gc collect

* fix hygiene, for the most part

* clean up

* go back to old span representation

* clean up

* fix build

* remove dummy source

+++ crates/steel-core/src/parser/span.rs
@@ -24,7 +24,7 @@ impl FromSteelVal for Span {
                 end: usize::from_steelval(l.get(1).unwrap())?,
                 source_id: l
                     .get(2)
-                    .map(Option::<usize>::from_steelval)
+                    .map(Option::<u32>::from_steelval)
                     .map(|x| x.transpose())
                     .flatten()
                     .transpose()?
